Added endpoint for VERY SIMPLE account banning and cleaned up some code

+++ Spring/src/main/java/pl/ttsw/GameRev/repository/GameRepository.java
@@ -7,5 +7,5 @@ import pl.ttsw.GameRev.model.Game;
 @Repository
 public interface GameRepository extends JpaRepository<Game, Integer> {
     Game findGameByTitle(String title);
-    Game findGameById(int id);
+    Game findGameById(Long id);
 }
